招新

+++ controllers/request.go
@@ -11,13 +11,13 @@ const CtxUserIDKey = "userID"
 
 var ErrorUserNotLogin = errors.New("用户未登录")
 
-func getCurrentUser(c *gin.Context) (userID int64, err error) {
+func getCurrentUser(c *gin.Context) (userID uint, err error) {
 	uid, ok := c.Get(CtxUserIDKey)
 	if !ok {
 		err = ErrorUserNotLogin
 		return
 	}
-	userID, ok = uid.(int64)
+	userID, ok = uid.(uint)
 	if !ok {
 		err = ErrorUserNotLogin
 		return
